Move non-provider related functions into a utils.go file
